connectors/qualys2tone/pyproject.toml
[project]
name = "qualys2tone"
version = "0.9.1"
description = "Qualys VM to Tenable One"
readme = "README.md"
authors = [
    {name = "Tenable, Inc", email = "ecosystem-support@tenable.com"},
]
requires-python = ">=3.12"
dependencies = [
    "arrow>=1.3.0",
    "lxml>=5.3.0",
    "pydantic-xml>=2.13.1",
    "pydantic>=2.9.2",
    "restfly>=1.5.0",
    "rich>=13.9.2",
    "sqlalchemy>=2.0.36",
    "typer>=0.12.5",
    "pytenable>=1.5.0",
    "pydantic-extra-types>=2.10.0",
    "tenint>=0.1.0",
    "defusedxml>=0.7.1",
]

[project.optional-dependencies]
testing = [
    "pytest-cov>=6.0.0",
    "responses>=0.25.3",
]

[project.urls]
logo = "https://raw.githubusercontent.com/tenable/tenable-connectors/refs/heads/main/connectors/qualys2tone/logo.svg"
support = "https://community.tenable.com"

[tool.tenint.connector]
title = "Qualys to Tenable One"
tags= ["qualys", "tvm"]
timeout = 3600

[tool.uv.sources]
pytenable = {url = "https://github.com/tenable/pyTenable/archive/refs/heads/feature/sync.zip"}

[tool.uv]
dev-dependencies = [
    "ptpython>=3.0.29",
    "pyright>=1.1.385",
    "pytest-cov>=5.0.0",
    "pytest>=8.3.3",
    "responses>=0.25.3",
    "ruff>=0.7.0",
    "isort>=5.13.2",
]

[tool.ruff]
line-length = 88
indent-width = 4
exclude = [
  ".nova",
  ".github",
  ".git",
  ".pytest_cache",
  "__pycache__"
]

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B"]
fixable = [ "ALL" ]
unfixable = [ "B" ]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
line-ending = "lf"
docstring-code-format = false
docstring-code-line-length = "dynamic"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
